requirements.txt
pyyaml
haikunator
azure